project/profile_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
email-validator==2.2.0
jinja2==3.1.4
python-multipart==0.0.6


